Use the gocd api library for all api calls
