+++ fetch/src/main/java/com/indix/gocd/s3fetch/FetchConfigEnum.java
@@ -1,0 +1,13 @@
+package com.indix.gocd.s3fetch;
+
+enum FetchConfigEnum {
+    REPO("Repo"),
+    PACKAGE("Package"),
+    DESTINATION("Destination");
+
+    private final String propertyValue;
+    FetchConfigEnum(String sth) {
+        propertyValue = sth;
+    }
+
+}
