+++ fetch/src/main/java/com/indix/gocd/s3fetch/FetchTaskView.java
@@ -1,0 +1,11 @@
+package com.indix.gocd.s3fetch;
+
+import io.jmnarloch.cd.go.plugin.api.view.AbstractTaskView;
+
+class FetchTaskView extends AbstractTaskView {
+
+    public FetchTaskView(String displayValue, String templatePath) {
+        super(displayValue, templatePath);
+    }
+
+}
